+++ Runtime/Enums/SingletonMode.cs
@@ -1,5 +1,8 @@
 ﻿namespace TheRealIronDuck.Ducktion.Enums
 {
+    /// <summary>
+    /// This enum is used to define if a service is a singleton or not.
+    /// </summary>
     public enum SingletonMode
     {
         Singleton,