+++ Tests/Editor/Container/ConfiguratorsTest.cs
@@ -1,6 +1,4 @@
-using System.Reflection;
 using NUnit.Framework;
-using TheRealIronDuck.Ducktion.Configurators;
 using TheRealIronDuck.Ducktion.Editor.Tests.Editor.Stubs;
 
 namespace TheRealIronDuck.Ducktion.Editor.Tests.Editor.Container