+++ Tests/Editor/Stubs/SecondServiceWithLogger.cs
@@ -1,0 +1,12 @@
+using TheRealIronDuck.Ducktion.Logging;
+
+namespace TheRealIronDuck.Ducktion.Editor.Tests.Editor.Stubs
+{
+    public class SecondServiceWithLogger
+    {
+        public SecondServiceWithLogger(DucktionLogger logger)
+        {
+            logger.Log(LogLevel.Debug, "Hello from SecondServiceWithLogger!");
+        }
+    }
+}